requirements.txt
torch==2.0.1
Pillow==10.0.0
torchvision==0.15.2
numpy==1.24.4
matplotlib==3.7.2
pathlib==1.0.1
scikit-learn==1.3.0
Flask==2.3.3 
jsonify==0.5
